requirements.txt
# CourseAlign API - Requirements
# Production-quality FastAPI backend for lecture-textbook alignment

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.17

# OpenAI
openai==1.54.0
httpx==0.27.2

# Document Processing
python-pptx==1.0.2
PyMuPDF==1.24.13
python-docx==1.1.2

# Vector Search
faiss-cpu==1.8.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
packaging==24.1
